Added more code to the muxgenerator class
Although this branch is for the implementation of the converter class there was
the need to add some more responsability to the mux generator in order to
separate the multiplexer from the converter.

Now the multiplexer is responsible for wrapping all the data from all the ports
in a single bottle. This way the converter only needs to apply the function to
one single data structure (which is the multiplexer exit).