Created needed elements on the config file
Now the final user has to specify how many mux there will be (even
if it is just one). Then each mux will be named according to its index
(like [mux1] and [mux2]).

For each mux the program will expect one converter so there will be
an equal number of mux and converter. Missing the empty converter
for when the user doesn't want to apply any kind of conversion to
the mutex result.