Completed support for multiple multiplexers
Finished the TODO list present on issue #3.

Both the mux and converter generators now have vectors as
attributes so they work more as managers. This way on
mux generator is able to generate more than one mux.

The same is happening with the converter which is able to support
a converter for each mux. Actually the main function
expects the number of converters to be the same as the number of muxes.
In case these number don't match the program won't be able to generate
the code.