Bottle creator generator almost finished
It is now possible to configure the messages to be sent.

The bottle is almost fully configurable. Missing the handling of
lists. The generator creates the needed structure for the lists
but still doesn't fill them with the info from the configuration files.

It is also missing the fill of the list with the mux result.